+++ src/read.c
@@ -204,6 +204,9 @@ typedef struct avifDecoderItem
     avifBool hasUnsupportedEssentialProperty; // If true, this item cites a property flagged as 'essential' that libavif doesn't support (yet). Ignore the item, if so.
     avifBool ipmaSeen;    // if true, this item already received a property association
     avifBool progressive; // if true, this item has progressive layers (a1lx), but does not select a specific layer (the layer_id value in lsel is set to 0xFFFF)
+#if defined(AVIF_ENABLE_EXPERIMENTAL_AVIR)
+    avifPixelFormat coniPixelFormat; // Set from the CondensedImageBox, if present (AVIF_PIXEL_FORMAT_NONE otherwise)
+#endif
 } avifDecoderItem;
 AVIF_ARRAY_DECLARE(avifDecoderItemArray, avifDecoderItem, item);
 
@@ -734,6 +737,12 @@ typedef struct avifMeta
     // AVIF, this should point at an image item containing color planes, and all other items
     // are ignored unless they refer to this item in some way (alpha plane, EXIF/XMP metadata).
     uint32_t primaryItemID;
+
+#if defined(AVIF_ENABLE_EXPERIMENTAL_AVIR)
+    // If true, the fields above were extracted from a CondensedImageBox. It imposes some
+    // constraints on its optional extendedMeta field, such as forbidden item IDs, properties etc.
+    avifBool fromConi;
+#endif
 } avifMeta;
 
 static void avifMetaDestroy(avifMeta * meta);
@@ -1170,6 +1179,19 @@ static avifResult avifDecoderItemValidateProperties(const avifDecoderItem * item
         }
     }
 
+#if defined(AVIF_ENABLE_EXPERIMENTAL_AVIR)
+    if ((item->coniPixelFormat != AVIF_PIXEL_FORMAT_NONE) &&
+        (item->coniPixelFormat != avifCodecConfigurationBoxGetFormat(&configProp->u.av1C))) {
+        avifDiagnosticsPrintf(diag,
+                              "Item ID %u format [%s] specified by coni box does not match %s property format [%s]",
+                              item->id,
+                              avifPixelFormatToString(item->coniPixelFormat),
+                              configPropName,
+                              avifPixelFormatToString(avifCodecConfigurationBoxGetFormat(&configProp->u.av1C)));
+        return AVIF_RESULT_BMFF_PARSE_FAILED;
+    }
+#endif // AVIF_ENABLE_EXPERIMENTAL_AVIR
+
     if (strictFlags & AVIF_STRICT_CLAP_VALID) {
         const avifProperty * clapProp = avifPropertyArrayFind(&item->properties, "clap");
         if (clapProp) {
@@ -1593,9 +1615,15 @@ static avifResult avifDecoderFindMetadata(avifDecoder * decoder, avifMeta * meta
 
             // Advance past Annex A.2.1's header
             BEGIN_STREAM(exifBoxStream, exifContents.data, exifContents.size, &decoder->diag, "Exif header");
-            uint32_t exifTiffHeaderOffset;
-            AVIF_CHECKERR(avifROStreamReadU32(&exifBoxStream, &exifTiffHeaderOffset),
-                          AVIF_RESULT_BMFF_PARSE_FAILED); // unsigned int(32) exif_tiff_header_offset;
+#if defined(AVIF_ENABLE_EXPERIMENTAL_AVIR)
+            // The CondensedImageBox does not signal the exifTiffHeaderOffset.
+            if (!meta->fromConi)
+#endif
+            {
+                uint32_t exifTiffHeaderOffset;
+                AVIF_CHECKERR(avifROStreamReadU32(&exifBoxStream, &exifTiffHeaderOffset),
+                              AVIF_RESULT_BMFF_PARSE_FAILED); // unsigned int(32) exif_tiff_header_offset;
+            }
 
             AVIF_CHECKRES(avifRWDataSet(&image->exif, avifROStreamCurrent(&exifBoxStream), avifROStreamRemainingBytes(&exifBoxStream)));
         } else if (!decoder->ignoreXMP && !memcmp(item->type, "mime", 4) &&
@@ -1691,6 +1719,14 @@ static avifBool avifParseItemLocationBox(avifMeta * meta, const uint8_t * raw, s
         } else {
             AVIF_CHECK(avifROStreamReadU32(&s, &itemID)); // unsigned int(32) item_ID;
         }
+#if defined(AVIF_ENABLE_EXPERIMENTAL_AVIR)
+        if (meta->fromConi && itemID < 5) {
+            // This is the extendedMeta field of an enclosing CondensedImageBox.
+            // Item IDs 1 to 4 are reserved and already defined, including their locations.
+            avifDiagnosticsPrintf(diag, "%s: Box[iloc] has a forbidden item ID [%u]", s.diagContext, itemID);
+            return AVIF_FALSE;
+        }
+#endif // AVIF_ENABLE_EXPERIMENTAL_AVIR
 
         avifDecoderItem * item = avifMetaFindItem(meta, itemID);
         if (!item) {
@@ -1717,6 +1753,14 @@ static avifBool avifParseItemLocationBox(avifMeta * meta, const uint8_t * raw, s
                 return AVIF_FALSE;
             }
             if (constructionMethod == 1) {
+#if defined(AVIF_ENABLE_EXPERIMENTAL_AVIR)
+                if (meta->fromConi) {
+                    // This is the extendedMeta field of an enclosing CondensedImageBox.
+                    // It may not contain an ItemDataBox.
+                    avifDiagnosticsPrintf(diag, "%s: Box[iloc] has a forbidden construction method [%u]", s.diagContext, constructionMethod);
+                    return AVIF_FALSE;
+                }
+#endif // AVIF_ENABLE_EXPERIMENTAL_AVIR
                 item->idatStored = AVIF_TRUE;
             }
         }
@@ -2182,6 +2226,14 @@ static avifBool avifParseItemPropertyContainerBox(avifPropertyArray * properties
 
 static avifBool avifParseItemPropertyAssociation(avifMeta * meta, const uint8_t * raw, size_t rawLen, avifDiagnostics * diag, uint32_t * outVersionAndFlags)
 {
+#if defined(AVIF_ENABLE_EXPERIMENTAL_AVIR)
+    if (meta->fromConi) {
+        // The CondensedImageBox will always create 8 item properties, so to refer to the
+        // first property in the ItemPropertyContainerBox of its extendedMeta field, use index 9.
+        assert(meta->properties.count >= 8);
+    }
+#endif // AVIF_ENABLE_EXPERIMENTAL_AVIR
+
     // NOTE: If this function ever adds support for versions other than [0,1] or flags other than
     //       [0,1], please increase the value of MAX_IPMA_VERSION_AND_FLAGS_SEEN accordingly.
 
@@ -2481,6 +2533,15 @@ static avifBool avifParseItemInfoEntry(avifMeta * meta, const uint8_t * raw, siz
         memset(&contentType, 0, sizeof(contentType));
     }
 
+#if defined(AVIF_ENABLE_EXPERIMENTAL_AVIR)
+    if (meta->fromConi && itemID < 5) {
+        // This is the extendedMeta field of an enclosing CondensedImageBox.
+        // itemID 1 is reserved for the primary color item, 2 for the alpha auxiliary item,
+        // 3 for the Exif metadata item and 4 for the XMP metadata item.
+        avifDiagnosticsPrintf(diag, "%s: Box[infe] of type %.4s has a forbidden item ID [%u]", s.diagContext, itemType, itemID);
+        return AVIF_FALSE;
+    }
+#endif // AVIF_ENABLE_EXPERIMENTAL_AVIR
     avifDecoderItem * item = avifMetaFindItem(meta, itemID);
     if (!item) {
         avifDiagnosticsPrintf(diag, "%s: Box[infe] with item_type %.4s has an invalid item_ID [%u]", s.diagContext, itemType, itemID);
@@ -3151,6 +3212,352 @@ static avifBool avifParseMovieBox(avifDecoderData * data, uint64_t rawOffset, co
     return AVIF_TRUE;
 }
 
+#if defined(AVIF_ENABLE_EXPERIMENTAL_AVIR)
+static avifBool avifParseExtendedMeta(avifMeta * meta, uint64_t rawOffset, const uint8_t * raw, size_t rawLen, avifDiagnostics * diag)
+{
+    BEGIN_STREAM(s, raw, rawLen, diag, "extendedMeta");
+    // The extendedMeta field has no size and box type because these are already set by the CondensedImageBox.
+
+    uint32_t uniqueBoxFlags = 0;
+    while (avifROStreamHasBytesLeft(&s, 1)) {
+        avifBoxHeader header;
+        AVIF_CHECK(avifROStreamReadBoxHeader(&s, &header));
+
+        if (!memcmp(header.type, "hdlr", 4) || !memcmp(header.type, "dinf", 4) || !memcmp(header.type, "idat", 4) ||
+            !memcmp(header.type, "pitm", 4)) {
+            avifDiagnosticsPrintf(diag, "Box[coni] shall have no Box[%.4s] in its extendedMeta field", (const char *)header.type);
+            return AVIF_FALSE;
+        } else if (!memcmp(header.type, "iinf", 4)) {
+            AVIF_CHECK(uniqueBoxSeen(&uniqueBoxFlags, 5, "coni", "iinf", diag));
+            AVIF_CHECK(avifParseItemInfoBox(meta, avifROStreamCurrent(&s), header.size, diag));
+        } else if (!memcmp(header.type, "iloc", 4)) {
+            AVIF_CHECK(uniqueBoxSeen(&uniqueBoxFlags, 1, "coni", "iloc", diag));
+            AVIF_CHECK(avifParseItemLocationBox(meta, avifROStreamCurrent(&s), header.size, diag));
+        } else if (!memcmp(header.type, "iprp", 4)) {
+            AVIF_CHECK(uniqueBoxSeen(&uniqueBoxFlags, 4, "coni", "iprp", diag));
+            AVIF_CHECK(avifParseItemPropertiesBox(meta, rawOffset + avifROStreamOffset(&s), avifROStreamCurrent(&s), header.size, diag));
+        } else if (!memcmp(header.type, "iref", 4)) {
+            AVIF_CHECK(uniqueBoxSeen(&uniqueBoxFlags, 6, "coni", "iref", diag));
+            AVIF_CHECK(avifParseItemReferenceBox(meta, avifROStreamCurrent(&s), header.size, diag));
+        }
+
+        AVIF_CHECK(avifROStreamSkip(&s, header.size));
+    }
+    return AVIF_TRUE;
+}
+
+static avifProperty * avifMetaCreateProperty(avifMeta * meta, const char * propertyType)
+{
+    avifProperty * metaProperty = avifArrayPushPtr(&meta->properties);
+    AVIF_CHECK(metaProperty);
+    memcpy(metaProperty->type, propertyType, 4);
+    return metaProperty;
+}
+
+static avifProperty * avifDecoderItemAddProperty(avifDecoderItem * item, const avifProperty * metaProperty)
+{
+    avifProperty * itemProperty = avifArrayPushPtr(&item->properties);
+    AVIF_CHECK(itemProperty);
+    *itemProperty = *metaProperty;
+    return itemProperty;
+}
+
+static avifResult avifParseCondensedImageBox(avifMeta * meta, uint64_t rawOffset, const uint8_t * raw, size_t rawLen, avifDiagnostics * diag)
+{
+    // Experimental box.
+    BEGIN_STREAM(s, raw, rawLen, diag, "Box[coni]");
+
+    meta->fromConi = AVIF_TRUE;
+
+    // Read the bit fields.
+    // TODO(yguyon): Implement or return an appropriate error in cases where AVIF_RESULT_NOT_IMPLEMENTED is returned.
+
+    uint32_t version;
+    AVIF_CHECKERR(avifROStreamReadBits(&s, &version, 2), AVIF_RESULT_BMFF_PARSE_FAILED); // unsigned int(2) version;
+    AVIF_CHECKERR(version == 0, AVIF_RESULT_NOT_IMPLEMENTED);
+
+    uint32_t width, height;
+    AVIF_CHECKERR(avifROStreamReadVarInt(&s, &width), AVIF_RESULT_BMFF_PARSE_FAILED);  // varint(32) width_minus_one;
+    AVIF_CHECKERR(avifROStreamReadVarInt(&s, &height), AVIF_RESULT_BMFF_PARSE_FAILED); // varint(32) height_minus_one;
+    ++width;
+    ++height;
+
+    uint32_t isFloat;
+    AVIF_CHECKERR(avifROStreamReadBits(&s, &isFloat, 1), AVIF_RESULT_BMFF_PARSE_FAILED); // unsigned int(1) is_float;
+    AVIF_CHECKERR(!isFloat, AVIF_RESULT_NOT_IMPLEMENTED);
+    uint32_t bitDepth;
+    AVIF_CHECKERR(avifROStreamReadBits(&s, &bitDepth, 4), AVIF_RESULT_BMFF_PARSE_FAILED); // unsigned int(1) bit_depth_minus_one;
+    ++bitDepth;
+    AVIF_CHECKERR((bitDepth == 8) || (bitDepth == 10) || (bitDepth == 12), AVIF_RESULT_UNSUPPORTED_DEPTH);
+    uint32_t isMonochrome;
+    AVIF_CHECKERR(avifROStreamReadBits(&s, &isMonochrome, 1), AVIF_RESULT_BMFF_PARSE_FAILED); // unsigned int(1) is_monochrome;
+    uint32_t isSubsampled = AVIF_FALSE;
+    if (!isMonochrome) {
+        AVIF_CHECKERR(avifROStreamReadBits(&s, &isSubsampled, 1), AVIF_RESULT_BMFF_PARSE_FAILED); // unsigned int(1) is_subsampled;
+    }
+    uint32_t fullRange;
+    AVIF_CHECKERR(avifROStreamReadBits(&s, &fullRange, 1), AVIF_RESULT_BMFF_PARSE_FAILED); // unsigned int(1) full_range;
+    uint32_t colorType;
+    AVIF_CHECKERR(avifROStreamReadBits(&s, &colorType, 2), AVIF_RESULT_BMFF_PARSE_FAILED); // unsigned int(2) color_type;
+    uint32_t colorPrimaries;
+    uint32_t transferCharacteristics;
+    uint32_t matrixCoefficients;
+    uint32_t iccDataSize;
+    if (colorType == AVIF_CONI_COLOR_TYPE_ICC) {
+        colorPrimaries = AVIF_COLOR_PRIMARIES_UNSPECIFIED;
+        transferCharacteristics = AVIF_TRANSFER_CHARACTERISTICS_UNSPECIFIED;
+        AVIF_CHECKERR(avifROStreamReadBits(&s, &matrixCoefficients, 8), AVIF_RESULT_BMFF_PARSE_FAILED); // unsigned int(8) matrix_coefficients;
+        AVIF_CHECKERR(avifROStreamReadVarInt(&s, &iccDataSize), AVIF_RESULT_BMFF_PARSE_FAILED); // varint(32) icc_data_size;
+        ++iccDataSize;
+    } else if (colorType == AVIF_CONI_COLOR_TYPE_SRGB) {
+        // sRGB
+        colorPrimaries = AVIF_COLOR_PRIMARIES_BT709;
+        transferCharacteristics = AVIF_TRANSFER_CHARACTERISTICS_SRGB;
+        matrixCoefficients = AVIF_MATRIX_COEFFICIENTS_BT601;
+        iccDataSize = 0;
+    } else {
+        const uint32_t numBitsPerComponent = (colorType == AVIF_CONI_COLOR_TYPE_NCLX_5BIT) ? 5 : 8;
+        AVIF_CHECKERR(avifROStreamReadBits(&s, &colorPrimaries, numBitsPerComponent),
+                      AVIF_RESULT_BMFF_PARSE_FAILED); // unsigned int(5/8) color_primaries;
+        AVIF_CHECKERR(avifROStreamReadBits(&s, &transferCharacteristics, numBitsPerComponent),
+                      AVIF_RESULT_BMFF_PARSE_FAILED); // unsigned int(5/8) transfer_characteristics;
+        AVIF_CHECKERR(avifROStreamReadBits(&s, &matrixCoefficients, numBitsPerComponent),
+                      AVIF_RESULT_BMFF_PARSE_FAILED); // unsigned int(5/8) matrix_coefficients;
+        iccDataSize = 0;
+    }
+
+    uint32_t hasExplicitCodecTypes;
+    AVIF_CHECKERR(avifROStreamReadBits(&s, &hasExplicitCodecTypes, 1), AVIF_RESULT_BMFF_PARSE_FAILED); // unsigned int(1) has_explicit_codec_types;
+    AVIF_CHECKERR(!hasExplicitCodecTypes, AVIF_RESULT_NOT_IMPLEMENTED);
+    uint32_t colorItemCodecConfigSize, colorItemDataSize;
+    AVIF_CHECKERR(avifROStreamReadVarInt(&s, &colorItemCodecConfigSize), AVIF_RESULT_BMFF_PARSE_FAILED); // varint(32) main_item_codec_config_size;
+    AVIF_CHECKERR(colorItemCodecConfigSize == 4, AVIF_RESULT_NOT_IMPLEMENTED);
+    AVIF_CHECKERR(avifROStreamReadVarInt(&s, &colorItemDataSize), AVIF_RESULT_BMFF_PARSE_FAILED); // varint(32) main_item_data_size;
+    ++colorItemDataSize;
+
+    uint32_t hasAlpha;
+    uint32_t alphaIsPremultiplied = AVIF_FALSE;
+    uint32_t alphaItemDataSize = 0;
+    AVIF_CHECKERR(avifROStreamReadBits(&s, &hasAlpha, 1), AVIF_RESULT_BMFF_PARSE_FAILED); // unsigned int(1) has_alpha;
+    if (hasAlpha) {
+        AVIF_CHECKERR(avifROStreamReadBits(&s, &alphaIsPremultiplied, 1), AVIF_RESULT_BMFF_PARSE_FAILED); // unsigned int(1) alpha_is_premultiplied;
+        uint32_t alphaItemCodecConfigSize;
+        AVIF_CHECKERR(avifROStreamReadVarInt(&s, &alphaItemCodecConfigSize), AVIF_RESULT_BMFF_PARSE_FAILED); // varint(32) alpha_item_codec_config_size;
+        AVIF_CHECKERR(alphaItemCodecConfigSize == 4, AVIF_RESULT_NOT_IMPLEMENTED);
+        AVIF_CHECKERR(avifROStreamReadVarInt(&s, &alphaItemDataSize), AVIF_RESULT_BMFF_PARSE_FAILED); // varint(32) alpha_item_data_size;
+    }
+
+    uint32_t hasExtendedMeta, hasExif, hasXMP;
+    uint32_t extendedMetaSize = 0, exifSize = 0, xmpSize = 0;
+    AVIF_CHECKERR(avifROStreamReadBits(&s, &hasExtendedMeta, 1), AVIF_RESULT_BMFF_PARSE_FAILED); // unsigned int(1) has_extended_meta;
+    if (hasExtendedMeta) {
+        AVIF_CHECKERR(avifROStreamReadVarInt(&s, &extendedMetaSize), AVIF_RESULT_BMFF_PARSE_FAILED); // varint(32) extended_meta_size;
+        ++extendedMetaSize;
+    }
+    AVIF_CHECKERR(avifROStreamReadBits(&s, &hasExif, 1), AVIF_RESULT_BMFF_PARSE_FAILED); // unsigned int(1) has_exif;
+    if (hasExif) {
+        AVIF_CHECKERR(avifROStreamReadVarInt(&s, &exifSize), AVIF_RESULT_BMFF_PARSE_FAILED); // varint(32) exif_data_size;
+        ++exifSize;
+    }
+    AVIF_CHECKERR(avifROStreamReadBits(&s, &hasXMP, 1), AVIF_RESULT_BMFF_PARSE_FAILED); // unsigned int(1) has_xmp;
+    if (hasXMP) {
+        AVIF_CHECKERR(avifROStreamReadVarInt(&s, &xmpSize), AVIF_RESULT_BMFF_PARSE_FAILED); // varint(32) xmp_data_size;
+        ++xmpSize;
+    }
+
+    // Padding to align with whole bytes if necessary.
+    if (s.numUsedBitsInPartialByte) {
+        uint32_t padding;
+        AVIF_CHECKERR(avifROStreamReadBits(&s, &padding, 8 - s.numUsedBitsInPartialByte), AVIF_RESULT_BMFF_PARSE_FAILED);
+        AVIF_CHECKERR(!padding, AVIF_RESULT_BMFF_PARSE_FAILED); // Only accept zeros as padding.
+    }
+
+    avifCodecConfigurationBox alphaCodecConfig = { 0 };
+    if (hasAlpha) {
+        AVIF_CHECKERR(avifParseCodecConfiguration(&s, &alphaCodecConfig, "coni", diag),
+                      AVIF_RESULT_BMFF_PARSE_FAILED); // unsigned int(8) alpha_item_codec_config[];
+    }
+    avifCodecConfigurationBox colorCodecConfig;
+    AVIF_CHECKERR(avifParseCodecConfiguration(&s, &colorCodecConfig, "coni", diag),
+                  AVIF_RESULT_BMFF_PARSE_FAILED); // unsigned int(8) main_item_codec_config[];
+
+    // Store and update the offset for the following item extents and properties.
+    // The extendedMeta field is parsed after creating the items defined by the CondensedImageBox
+    // so the stream s cannot be used for keeping track of the position.
+    size_t offset = (size_t)rawOffset + avifROStreamOffset(&s) + (size_t)extendedMetaSize;
+
+    // Create the items and properties generated by the CondensedImageBox.
+    // The CondensedImageBox always creates 8 properties for specification easiness.
+    // Use FreeSpaceBoxes as no-op placeholder properties when necessary.
+    // There is no need to use placeholder items because item IDs do not have to
+    // be contiguous, whereas property indices shall be 1, 2, 3, 4, 5 etc.
+
+    meta->primaryItemID = 1;
+    avifDecoderItem * colorItem = avifMetaFindItem(meta, meta->primaryItemID);
+    AVIF_CHECKERR(colorItem, AVIF_RESULT_OUT_OF_MEMORY);
+    memcpy(colorItem->type, "av01", 4);
+    colorItem->width = width;
+    colorItem->height = height;
+    colorItem->coniPixelFormat = isMonochrome   ? AVIF_PIXEL_FORMAT_YUV400
+                                 : isSubsampled ? AVIF_PIXEL_FORMAT_YUV420
+                                                : AVIF_PIXEL_FORMAT_YUV444;
+
+    avifDecoderItem * alphaItem = NULL;
+    if (hasAlpha) {
+        alphaItem = avifMetaFindItem(meta, /*itemID=*/2);
+        AVIF_CHECKERR(alphaItem, AVIF_RESULT_OUT_OF_MEMORY);
+        memcpy(alphaItem->type, "av01", 4);
+        alphaItem->width = width;
+        alphaItem->height = height;
+        alphaItem->coniPixelFormat = AVIF_PIXEL_FORMAT_YUV400;
+    }
+
+    // Property with fixed index 1.
+    avifProperty * colorCodecConfigProp = avifMetaCreateProperty(meta, "av1C");
+    AVIF_CHECKERR(colorCodecConfigProp, AVIF_RESULT_OUT_OF_MEMORY);
+    colorCodecConfigProp->u.av1C = colorCodecConfig;
+    AVIF_CHECKERR(avifDecoderItemAddProperty(colorItem, colorCodecConfigProp), AVIF_RESULT_OUT_OF_MEMORY);
+
+    // Property with fixed index 2.
+    avifProperty * ispeProp = avifMetaCreateProperty(meta, "ispe");
+    AVIF_CHECKERR(ispeProp, AVIF_RESULT_OUT_OF_MEMORY);
+    ispeProp->u.ispe.width = width;
+    ispeProp->u.ispe.height = height;
+    AVIF_CHECKERR(avifDecoderItemAddProperty(colorItem, ispeProp), AVIF_RESULT_OUT_OF_MEMORY);
+
+    // Property with fixed index 3.
+    avifProperty * pixiProp = avifMetaCreateProperty(meta, "pixi");
+    AVIF_CHECKERR(pixiProp, AVIF_RESULT_OUT_OF_MEMORY);
+    pixiProp->u.pixi.planeCount = isMonochrome ? 1 : 3;
+    for (uint8_t plane = 0; plane < pixiProp->u.pixi.planeCount; ++plane) {
+        pixiProp->u.pixi.planeDepths[plane] = (uint8_t)bitDepth;
+    }
+    AVIF_CHECKERR(avifDecoderItemAddProperty(colorItem, pixiProp), AVIF_RESULT_OUT_OF_MEMORY);
+
+    // Property with fixed index 4.
+    avifProperty * colrPropNCLX = avifMetaCreateProperty(meta, "colr");
+    AVIF_CHECKERR(colrPropNCLX, AVIF_RESULT_OUT_OF_MEMORY);
+    colrPropNCLX->u.colr.hasNCLX = AVIF_TRUE; // colour_type "nclx"
+    colrPropNCLX->u.colr.colorPrimaries = (avifColorPrimaries)colorPrimaries;
+    colrPropNCLX->u.colr.transferCharacteristics = (avifTransferCharacteristics)transferCharacteristics;
+    colrPropNCLX->u.colr.matrixCoefficients = (avifMatrixCoefficients)matrixCoefficients;
+    colrPropNCLX->u.colr.range = fullRange ? AVIF_RANGE_FULL : AVIF_RANGE_LIMITED;
+    AVIF_CHECKERR(avifDecoderItemAddProperty(colorItem, colrPropNCLX), AVIF_RESULT_OUT_OF_MEMORY);
+
+    // Property with fixed index 5.
+    if (iccDataSize) {
+        avifProperty * colrPropICC = avifMetaCreateProperty(meta, "colr");
+        AVIF_CHECKERR(colrPropICC, AVIF_RESULT_OUT_OF_MEMORY);
+        colrPropICC->u.colr.hasICC = AVIF_TRUE; // colour_type "rICC" or "prof"
+        colrPropICC->u.colr.iccOffset = offset;
+        colrPropICC->u.colr.iccSize = iccDataSize;
+        offset += iccDataSize;
+        AVIF_CHECKERR(avifDecoderItemAddProperty(colorItem, colrPropICC), AVIF_RESULT_OUT_OF_MEMORY);
+    } else {
+        AVIF_CHECKERR(avifMetaCreateProperty(meta, "skip"), AVIF_RESULT_OUT_OF_MEMORY); // Placeholder.
+    }
+
+    if (hasAlpha) {
+        // Property with fixed index 6.
+        avifProperty * alphaCodecConfigProp = avifMetaCreateProperty(meta, "av1C");
+        AVIF_CHECKERR(alphaCodecConfigProp, AVIF_RESULT_OUT_OF_MEMORY);
+        alphaCodecConfigProp->u.av1C = alphaCodecConfig;
+        AVIF_CHECKERR(avifDecoderItemAddProperty(alphaItem, alphaCodecConfigProp), AVIF_RESULT_OUT_OF_MEMORY);
+
+        // Property with fixed index 7.
+        alphaItem->auxForID = colorItem->id;
+        colorItem->premByID = alphaIsPremultiplied;
+        avifProperty * alphaAuxProp = avifMetaCreateProperty(meta, "auxC");
+        AVIF_CHECKERR(alphaAuxProp, AVIF_RESULT_OUT_OF_MEMORY);
+        strcpy(alphaAuxProp->u.auxC.auxType, AVIF_URN_ALPHA0);
+        AVIF_CHECKERR(avifDecoderItemAddProperty(alphaItem, alphaAuxProp), AVIF_RESULT_OUT_OF_MEMORY);
+
+        // Property with fixed index 2 (reused).
+        AVIF_CHECKERR(avifDecoderItemAddProperty(alphaItem, ispeProp), AVIF_RESULT_OUT_OF_MEMORY);
+
+        // Property with fixed index 8.
+        avifProperty * alphaPixiProp = avifMetaCreateProperty(meta, "pixi");
+        AVIF_CHECKERR(alphaPixiProp, AVIF_RESULT_OUT_OF_MEMORY);
+        memcpy(alphaPixiProp->type, "pixi", 4);
+        alphaPixiProp->u.pixi.planeCount = 1;
+        alphaPixiProp->u.pixi.planeDepths[0] = (uint8_t)bitDepth;
+        AVIF_CHECKERR(avifDecoderItemAddProperty(alphaItem, alphaPixiProp), AVIF_RESULT_OUT_OF_MEMORY);
+    } else {
+        // Placeholders 6, 7 and 8.
+        AVIF_CHECKERR(avifMetaCreateProperty(meta, "skip"), AVIF_RESULT_OUT_OF_MEMORY);
+        AVIF_CHECKERR(avifMetaCreateProperty(meta, "skip"), AVIF_RESULT_OUT_OF_MEMORY);
+        AVIF_CHECKERR(avifMetaCreateProperty(meta, "skip"), AVIF_RESULT_OUT_OF_MEMORY);
+    }
+
+    // Extents.
+
+    if (hasAlpha) {
+        avifExtent * alphaExtent = (avifExtent *)avifArrayPushPtr(&alphaItem->extents);
+        AVIF_CHECKERR(alphaExtent, AVIF_RESULT_OUT_OF_MEMORY);
+        alphaExtent->offset = offset;
+        alphaExtent->size = alphaItemDataSize;
+        offset += alphaItemDataSize;
+        alphaItem->size = alphaItemDataSize;
+    }
+
+    avifExtent * colorExtent = (avifExtent *)avifArrayPushPtr(&colorItem->extents);
+    AVIF_CHECKERR(colorExtent, AVIF_RESULT_OUT_OF_MEMORY);
+    colorExtent->offset = offset;
+    colorExtent->size = colorItemDataSize;
+    offset += colorItemDataSize;
+    colorItem->size = colorItemDataSize;
+
+    if (hasExif) {
+        avifDecoderItem * exifItem = avifMetaFindItem(meta, /*itemID=*/3);
+        AVIF_CHECKERR(exifItem, AVIF_RESULT_OUT_OF_MEMORY);
+        memcpy(exifItem->type, "Exif", 4);
+        exifItem->descForID = colorItem->id;
+        colorItem->premByID = alphaIsPremultiplied;
+
+        avifExtent * exifExtent = (avifExtent *)avifArrayPushPtr(&exifItem->extents);
+        AVIF_CHECKERR(exifExtent, AVIF_RESULT_OUT_OF_MEMORY);
+        exifExtent->offset = offset;
+        exifExtent->size = exifSize; // Does not include unsigned int(32) exif_tiff_header_offset;
+        offset += exifSize;
+        exifItem->size = exifExtent->size;
+    }
+
+    if (hasXMP) {
+        avifDecoderItem * xmpItem = avifMetaFindItem(meta, /*itemID=*/4);
+        AVIF_CHECKERR(xmpItem, AVIF_RESULT_OUT_OF_MEMORY);
+        memcpy(xmpItem->type, "mime", 4);
+        memcpy(xmpItem->contentType.contentType, xmpContentType, xmpContentTypeSize);
+        xmpItem->descForID = colorItem->id;
+        colorItem->premByID = alphaIsPremultiplied;
+
+        avifExtent * xmpExtent = (avifExtent *)avifArrayPushPtr(&xmpItem->extents);
+        AVIF_CHECKERR(xmpExtent, AVIF_RESULT_OUT_OF_MEMORY);
+        xmpExtent->offset = offset;
+        xmpExtent->size = xmpSize;
+        offset += xmpSize;
+        xmpItem->size = xmpSize;
+    }
+
+    // Complete the generated virtual MetaBox with the ExtendedMetaBox items and properties.
+    // The ExtendedMetaBox may reuse items and properties created above so it must be parsed last.
+
+    if (hasExtendedMeta) {
+        AVIF_CHECKERR(avifParseExtendedMeta(meta, rawOffset + avifROStreamOffset(&s), avifROStreamCurrent(&s), extendedMetaSize, diag),
+                      AVIF_RESULT_BMFF_PARSE_FAILED);
+    }
+
+    // Check that the stream s and the variable offset match.
+    AVIF_CHECKERR(avifROStreamSkip(&s, extendedMetaSize), AVIF_RESULT_TRUNCATED_DATA);
+    AVIF_CHECKERR(avifROStreamSkip(&s, iccDataSize), AVIF_RESULT_TRUNCATED_DATA);
+    AVIF_CHECKERR(avifROStreamSkip(&s, alphaItemDataSize), AVIF_RESULT_TRUNCATED_DATA);
+    AVIF_CHECKERR(avifROStreamSkip(&s, colorItemDataSize), AVIF_RESULT_TRUNCATED_DATA);
+    AVIF_CHECKERR(avifROStreamSkip(&s, exifSize), AVIF_RESULT_TRUNCATED_DATA);
+    AVIF_CHECKERR(avifROStreamSkip(&s, xmpSize), AVIF_RESULT_TRUNCATED_DATA);
+    assert(rawOffset + avifROStreamOffset(&s) == offset);
+    return AVIF_RESULT_OK;
+}
+#endif // AVIF_ENABLE_EXPERIMENTAL_AVIR
+
 static avifBool avifParseFileTypeBox(avifFileType * ftyp, const uint8_t * raw, size_t rawLen, avifDiagnostics * diag)
 {
     BEGIN_STREAM(s, raw, rawLen, diag, "Box[ftyp]");
@@ -3186,6 +3593,10 @@ static avifResult avifParse(avifDecoder * decoder)
     avifBool moovSeen = AVIF_FALSE;
     avifBool needsMeta = AVIF_FALSE;
     avifBool needsMoov = AVIF_FALSE;
+#if defined(AVIF_ENABLE_EXPERIMENTAL_AVIR)
+    avifBool coniSeen = AVIF_FALSE;
+    avifBool needsConi = AVIF_FALSE;
+#endif
 
     for (;;) {
         // Read just enough to get the next box header (a max of 32 bytes)
@@ -3215,7 +3626,12 @@ static avifResult avifParse(avifDecoder * decoder)
         avifROData boxContents = AVIF_DATA_EMPTY;
 
         // TODO: reorg this code to only do these memcmps once each
+#if defined(AVIF_ENABLE_EXPERIMENTAL_AVIR)
+        if (!memcmp(header.type, "ftyp", 4) || !memcmp(header.type, "meta", 4) || !memcmp(header.type, "moov", 4) ||
+            !memcmp(header.type, "coni", 4)) {
+#else
         if (!memcmp(header.type, "ftyp", 4) || !memcmp(header.type, "meta", 4) || !memcmp(header.type, "moov", 4)) {
+#endif
             boxOffset = parseOffset;
             readResult = decoder->io->read(decoder->io, 0, parseOffset, header.size, &boxContents);
             if (readResult != AVIF_RESULT_OK) {
@@ -3241,6 +3657,12 @@ static avifResult avifParse(avifDecoder * decoder)
             memcpy(data->majorBrand, ftyp.majorBrand, 4); // Remember the major brand for future AVIF_DECODER_SOURCE_AUTO decisions
             needsMeta = avifFileTypeHasBrand(&ftyp, "avif");
             needsMoov = avifFileTypeHasBrand(&ftyp, "avis");
+#if defined(AVIF_ENABLE_EXPERIMENTAL_AVIR)
+            needsConi = avifFileTypeHasBrand(&ftyp, "avir");
+            if (needsConi && (needsMeta || needsMoov)) {
+                return AVIF_RESULT_INVALID_FTYP;
+            }
+#endif // AVIF_ENABLE_EXPERIMENTAL_AVIR
         } else if (!memcmp(header.type, "meta", 4)) {
             AVIF_CHECKERR(!metaSeen, AVIF_RESULT_BMFF_PARSE_FAILED);
             AVIF_CHECKERR(avifParseMetaBox(data->meta, boxOffset, boxContents.data, boxContents.size, data->diag),
@@ -3251,12 +3673,30 @@ static avifResult avifParse(avifDecoder * decoder)
             AVIF_CHECKERR(avifParseMovieBox(data, boxOffset, boxContents.data, boxContents.size, decoder->imageSizeLimit, decoder->imageDimensionLimit),
                           AVIF_RESULT_BMFF_PARSE_FAILED);
             moovSeen = AVIF_TRUE;
+#if defined(AVIF_ENABLE_EXPERIMENTAL_AVIR)
+        } else if (!memcmp(header.type, "coni", 4)) {
+            AVIF_CHECKERR(!metaSeen && !moovSeen, AVIF_RESULT_BMFF_PARSE_FAILED);
+            AVIF_CHECKRES(avifParseCondensedImageBox(data->meta, boxOffset, boxContents.data, boxContents.size, data->diag));
+            coniSeen = AVIF_TRUE;
+#endif // AVIF_ENABLE_EXPERIMENTAL_AVIR
         }
+
+#if defined(AVIF_ENABLE_EXPERIMENTAL_AVIR)
+        if (ftypSeen && !needsConi && coniSeen) {
+            // The 'coni' box should be ignored if there is no 'avir' brand, but libavif allows reading them in any order.
+            return AVIF_RESULT_NOT_IMPLEMENTED; // TODO(yguyon): Implement
+        }
+#endif // AVIF_ENABLE_EXPERIMENTAL_AVIR
 
         // See if there is enough information to consider Parse() a success and early-out:
         // * If the brand 'avif' is present, require a meta box
         // * If the brand 'avis' is present, require a moov box
+        // * If AVIF_ENABLE_EXPERIMENTAL_AVIR is defined and the brand 'avir' is present, require a coni box
+#if defined(AVIF_ENABLE_EXPERIMENTAL_AVIR)
+        if (ftypSeen && (!needsMeta || metaSeen) && (!needsMoov || moovSeen) && (!needsConi || coniSeen)) {
+#else
         if (ftypSeen && (!needsMeta || metaSeen) && (!needsMoov || moovSeen)) {
+#endif
             return AVIF_RESULT_OK;
         }
     }
@@ -3266,6 +3706,11 @@ static avifResult avifParse(avifDecoder * decoder)
     if ((needsMeta && !metaSeen) || (needsMoov && !moovSeen)) {
         return AVIF_RESULT_TRUNCATED_DATA;
     }
+#if defined(AVIF_ENABLE_EXPERIMENTAL_AVIR)
+    if (needsConi && !coniSeen) {
+        return AVIF_RESULT_TRUNCATED_DATA;
+    }
+#endif
     return AVIF_RESULT_OK;
 }
 
@@ -3288,7 +3733,11 @@ static avifBool avifFileTypeHasBrand(avifFileType * ftyp, const char * brand)
 
 static avifBool avifFileTypeIsCompatible(avifFileType * ftyp)
 {
-    return avifFileTypeHasBrand(ftyp, "avif") || avifFileTypeHasBrand(ftyp, "avis");
+    return avifFileTypeHasBrand(ftyp, "avif") || avifFileTypeHasBrand(ftyp, "avis")
+#if defined(AVIF_ENABLE_EXPERIMENTAL_AVIR)
+           || avifFileTypeHasBrand(ftyp, "avir")
+#endif // AVIF_ENABLE_EXPERIMENTAL_AVIR
+        ;
 }
 
 avifBool avifPeekCompatibleFileType(const avifROData * input)